requirements.txt
Django==1.10.2
django-all-access==0.8.0
djangorestframework==3.4.7
funcsigs==1.0.2
Markdown==2.6.7
mock==2.0.0
oauthlib==2.0.0
pbr==1.10.0
pycrypto==2.6.1
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
